backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
paho-mqtt>=1.6.1
websockets>=10.1